services/auth/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
passlib
bcrypt==3.2.2
python-jose[cryptography]
python-multipart
email-validator
fastapi-mail
google-auth
requests
